requirements.txt
Django==3.0
mysqlclient==1.4.0
gunicorn==20.0.4
djangorestframework==3.12.2
drf-yasg==1.20.0
djangorestframework-simplejwt==4.6.0
pytest-django==4.1.0
pytest==6.2.1
